+++ hopsworksai/resource_cluster_test.go
@@ -1855,6 +1855,7 @@ func TestClusterUpdate(t *testing.T) {
 							"name": "cluster-name-1",
 							"state" : "running",
 							"provider": "AZURE",
+							"version": "v1",
 							"clusterConfiguration": {
 								"head": {
 									"instanceType": "node-type-1",
@@ -1878,6 +1879,7 @@ func TestClusterUpdate(t *testing.T) {
 		Id:                   "cluster-id-1",
 		Update:               true,
 		State: map[string]interface{}{
+			"version": "v1",
 			"workers": []interface{}{
 				map[string]interface{}{
 					"instance_type": "node-type-2",
@@ -2468,4 +2470,128 @@ func testGetAutoscaleConfig(reqBody io.Reader, cloud api.CloudProvider) (*api.Au
 		output = req.CreateRequest.Autoscale
 	}
 	return output, nil
+}
+
+func TestClusterUpdate_upgrade(t *testing.T) {
+	t.Parallel()
+	r := test.ResourceFixture{
+		HttpOps: []test.Operation{
+			{
+				Method: http.MethodGet,
+				Path:   "/api/clusters/cluster-id-1",
+				Response: `{
+					"apiVersion": "v1",
+					"status": "ok",
+					"code": 200,
+					"payload":{
+						"cluster": {
+							"id": "cluster-id-1",
+							"name": "cluster-name-1",
+							"state" : "running",
+							"provider": "AZURE",
+							"version": "v1",
+							"ports":{
+								"featureStore": false,
+								"onlineFeatureStore": false,
+								"kafka": false,
+								"ssh": false
+							}
+						}
+					}
+				}`,
+			},
+			{
+				Method: http.MethodPost,
+				Path:   "/api/clusters/cluster-id-1/upgrade",
+				ExpectRequestBody: `{
+					"version": "v2"
+				}`,
+				Response: `{
+					"apiVersion": "v1",
+					"status": "ok",
+					"code": 200
+				}`,
+			},
+		},
+		Resource:             clusterResource(),
+		OperationContextFunc: clusterResource().UpdateContext,
+		Id:                   "cluster-id-1",
+		Update:               true,
+		State: map[string]interface{}{
+			"version": "v2",
+			"open_ports": []interface{}{
+				map[string]interface{}{
+					"ssh":                  false,
+					"kafka":                false,
+					"feature_store":        false,
+					"online_feature_store": false,
+				},
+			},
+		},
+	}
+	r.Apply(t, context.TODO())
+}
+
+func TestClusterUpdate_upgrade_rollback(t *testing.T) {
+	t.Parallel()
+	r := test.ResourceFixture{
+		HttpOps: []test.Operation{
+			{
+				Method: http.MethodGet,
+				Path:   "/api/clusters/cluster-id-1",
+				Response: `{
+					"apiVersion": "v1",
+					"status": "ok",
+					"code": 200,
+					"payload":{
+						"cluster": {
+							"id": "cluster-id-1",
+							"name": "cluster-name-1",
+							"state" : "stopped",
+							"provider": "AZURE",
+							"version": "v2",
+							"ports":{
+								"featureStore": false,
+								"onlineFeatureStore": false,
+								"kafka": false,
+								"ssh": false
+							}
+						}
+					}
+				}`,
+			},
+			{
+				Method: http.MethodPut,
+				Path:   "/api/clusters/cluster-id-1/upgrade/rollback",
+				Response: `{
+					"apiVersion": "v1",
+					"status": "ok",
+					"code": 200
+				}`,
+			},
+		},
+		Resource:             clusterResource(),
+		OperationContextFunc: clusterResource().UpdateContext,
+		Id:                   "cluster-id-1",
+		Update:               true,
+		State: map[string]interface{}{
+			"version": "v1",
+			"state":   api.Error.String(),
+			"upgrade_in_progress": []interface{}{
+				map[string]interface{}{
+					"from_version": "v1",
+					"to_version":   "v2",
+				},
+			},
+			"open_ports": []interface{}{
+				map[string]interface{}{
+					"ssh":                  false,
+					"kafka":                false,
+					"feature_store":        false,
+					"online_feature_store": false,
+				},
+			},
+		},
+	}
+	r.Apply(t, context.TODO())
 }